+++ packages/chat/src/components/NewFooter/index.tsx
@@ -11,9 +11,10 @@ import { promptContainer } from '../Prompt/styles.css';
 import { ScrollToBottom } from '../ScrollToBottom';
 import {
   buttonsContainer,
-  extraLinkStyles,
+  externalLink,
   footerContainer,
   footerLinksContainer,
+  footerNote,
   hasEnded,
   messageContainer,
   poweredByLink,
@@ -51,7 +52,6 @@ export const NewFooter: React.FC<INewFooter> = ({
   scrollableAreaRef,
   onStart,
 }) => {
-  const showExtraLink = extraLinkText && extraLinkUrl;
   return (
     <div className={clsx(ClassName.FOOTER, footerContainer)}>
       <div className={promptContainer({ visible: messageInputProps.hasEnded })}>
@@ -84,9 +84,10 @@ export const NewFooter: React.FC<INewFooter> = ({
               </a>
             </div>
           )}
-          {showPoweredBy && showExtraLink && <div className={separator} />}
-          {showExtraLink && (
-            <a href={extraLinkUrl} target="_blank" className={extraLinkStyles}>
+          {showPoweredBy && extraLinkText && <div className={separator} />}
+          {extraLinkText && !extraLinkUrl && <div className={footerNote}>{extraLinkText}</div>}
+          {extraLinkText && extraLinkUrl && (
+            <a href={extraLinkUrl} target="_blank" className={externalLink}>
               {extraLinkText}
             </a>
           )}